Add refresh token.

+++ src/pages/Home.js
@@ -9,7 +9,7 @@ const Home = () => {
 
     const [userType, setUserType] = useState();
 
-    React.useEffect(() => {
+    useEffect(() => {
         axios.defaults.headers.common[
             "Authorization"
           ] = `Bearer ${localStorage.getItem(["access_token"])}`;
@@ -21,13 +21,39 @@ const Home = () => {
             setUserType(response.data["userType"])
         })
         .catch((error) => {
-            console.log(error);
-            localStorage.clear();
-            window.location.replace("/login");
-        })},
-        []
+            //리프레시 토큰 제출
+            const rData = new FormData();
+            rData.append("refresh_token", localStorage.getItem("refresh_token"))
+            console.log(rData);
+            axios
+            .post("http://localhost:8000/refreshToken",rData)
+            .then(response=>{
+                if (response.status === 200) {
+                    axios.defaults.headers.common[
+                    "Authorization"
+                    ] = `Bearer ${response.data["access_token"]}`;
+
+                    
+                    localStorage.setItem("access_token", response.data["access_token"])
+                    localStorage.setItem("token_type", response.data["token_type"])
+                    localStorage.setItem("refresh_token", response.data["refresh_token"])
+                    window.location.replace("/");
+            }
+            })
+            .catch(error => {
+            if(error.request.status===401){
+                console.log(error);
+                localStorage.clear();
+                window.location.replace("/login");
+            }else{
+                alert("System Error\nhome-1");
+            }
+            });
+        })},[]
     
-    )
+    );
+
+
